requirements.txt
feedparser==6.0.11
Flask==3.1.0
openai==1.68.2
python-dotenv==1.0.1
PyYAML==6.0.2
tenacity==9.0.0
tweepy==4.15.0
pypdf==5.4.0
gunicorn==23.0.0
